Cleaned up config file format
Removed the ability to specify flags on the command line. The only way of getting configuration into the router is now via the config files.
One can now specify two config files. The second config is overlayed onto the first one. This allows us to ship a standard config for all clients, and and then split off the client-specific stuff into isolated files.

Refactored the tests, to remove lots of repeated boiler-plate code.

The only new feature here is the ability to match arbitrary regexes on URLs.