Add HTTPS support
One other significant change here is to add the usage of a helper called tcpKeepAliveListener. This is behaviour that is copied from the Go standard library. It looks like it sets a property on the TCP port, that it must send TCP level keep-alives every three minutes. I assume that without this, the browser's connection would time out after a few minutes, if there were no requests made.