Simplify unit tests
The unit tests were written under the illusion that the Go unit testing
framework will launch multiple tests from the same OS process. This is not
true. The framework launches a new process for each test. This makes our
lives much simpler, since we don't have to worry about stopping our HTTP
server at the end of each test, which also happens to be impossible to do
in combination with HTTP/2.