TG-745; Change config loading from file to service
This changes the entire configuration loading from file mechanism.
Previously the main config and auxilary config files were specified on the
command line, this is not the case anymore. The merged config is provided
by the configuration service. The ability to load config from file is
retained but only for testing a simple config. The merging of main and aux
has been removed.