Fix bug in httpbridge requester forwarding
It was a mistake to create a new request object, and copy over some fields
from the original request. In doing so, I forgot to copy the ContentLength
field. In order to avoid any further such bugs, I'm now mutating the
original request object instead, which I hope will be more robust.