Add X-Original-Path header, and get rid of ms_http fork
The X-Original-Path header is for service-to-service authentication,
so that the receiver can validate the URL against the original URL
that the sender signed.
When service-to-service auth happens via the IMQS Router, then the
Path will often get altered (eg the service name is removed from
the prefix of the path). So it's vital for the receiving service
to have access to the original Path that was used to sign the message.
This is very much like the X-Forward-For header.

Also, got rid of the ms_http repo that we've been using for many
years. This was for the sake of Pure Technologies, who ran some old
IIS proxy servers.